Add galaxy
