+++ vintergatan-rust/src/main.rs
@@ -1,6 +1,14 @@
-use vintergatan_rust::model::position::Position;
+use vintergatan_rust::model::universe::Universe;
 
 fn main() {
-    let p = Position::new(0, 0);
-    println!("{}", p);
+    let width = 3;
+    let height = width;
+    let mut universe = Universe::new(width, height);
+    for n in 0..width*height*10 {
+        if n % 1 == 0 {
+            println!("{}score: {}\n", universe, universe.get_score());
+        }
+        universe.generate_step();
+    }
+    println!("{}", universe);
 }
