+++ vintergatan-rust/src/model/mod.rs
@@ -1,1 +1,3 @@
 pub mod position;
+pub mod galaxy;
+pub mod universe;
